requirements.txt
gunicorn==20.0.4
numpy==1.19.3
Pillow==8.0.1
plotly==4.13.0
pycairo==1.20.0
CairoSVG==2.5.0
dash==1.17.0
dash-bootstrap-components==0.10.7
tflite-runtime @ https://github.com/google-coral/pycoral/releases/download/release-frogfish/tflite_runtime-2.5.0-cp38-cp38-linux_x86_64.whl
Brotli==1.0.9                 # for dash
click==7.1.2                  # for dash
dash-core-components==1.13.0  # for dash
dash-html-components==1.1.1   # for dash
dash-renderer==1.8.3          # for dash
dash-table==4.11.0            # for dash
Flask==1.1.2                  # for dash
Flask-Compress==1.8.0         # for dash
future==0.18.2                # for dash
itsdangerous==1.1.0           # for dash
Jinja2==2.11.2                # for dash
MarkupSafe==1.1.1             # for dash
retrying==1.3.3               # for dash
six==1.15.0                   # for dash
Werkzeug==1.0.1               # for dash
cairocffi==1.2.0              # for CairoSVG
cffi==1.14.4                  # for CairoSVG
cssselect2==0.4.1             # for CairoSVG
defusedxml==0.6.0             # for CairoSVG
pycparser==2.20               # for CairoSVG
tinycss2==1.1.0               # for CairoSVG
webencodings==0.5.1           # for CairoSVG